refactor: extract components of `eventFormPopup`
